+++ NanoCETPy/models/arduino.py
@@ -2,13 +2,12 @@
     Modified Arduino model to accommodate peculiarities of NanoCET operation
 
 """
-
+from multiprocessing import Event
+from threading import RLock
+from experimentor.models.devices.base_device import ModelDevice
 import time
-
 import pyvisa
 from pyvisa import VisaIOError
-
-from NanoCETPy.dispertech.models.arduino import ArduinoModel
 from experimentor.lib.log import get_logger
 from experimentor.models import Feature
 from experimentor.models.decorators import make_async_thread
@@ -16,7 +15,7 @@ from experimentor.models.decorators import make_async_thread
 rm = pyvisa.ResourceManager('@py')
 
 
-class ArduinoNanoCET(ArduinoModel):
+class ArduinoNanoCET(ModelDevice):
     """
     ArduinoModel with modified initialize routine to enable NanoCET software connection check screen
 
@@ -29,8 +28,21 @@ class ArduinoNanoCET(ArduinoModel):
     Additional getters and setters for laser and LEDs have been added as the query string was changed in the Arduino firmware.
     """
     def __init__(self, port=None, device=0, baud_rate=9600, initial_config=None):
-        super().__init__(port=port, device=device, baud_rate=baud_rate, initial_config=initial_config)
+        """ Use the port if you know where the Arduino is connected, or use the device number in the order shown by
+        pyvisa.
+        """
+        super().__init__()
         self.logger = get_logger(__name__)
+        self._threads = []
+        self._stop_temperature = Event()
+        self.temp_electronics = 0
+        self.temp_sample = 0
+        self.query_lock = RLock()
+        self.driver = None
+        self.port = port
+        self.device = device
+        self.initial_config = initial_config
+        self.baud_rate = baud_rate
         self.initialized = False
         self.initializing = False
         self.led_states = {
@@ -51,10 +63,9 @@ class ArduinoNanoCET(ArduinoModel):
         self._top_led = 0
         self._fiber_led = 0
         self._side_led = 0
-        self._power_led = 0
+        self._power_led = 2  # Initial state set in the arduino code
         self._sample_led = 0
         self._measuring_led = 0
-        self.driver = None
 
     def serial_number(self):
         with self.query_lock:
@@ -281,10 +292,19 @@ class ArduinoNanoCET(ArduinoModel):
             return self.driver.query(f'LID').strip()
 
     def finalize(self):
+        self.logger.info('Finalizing Arduino')
+        if self.initial_config is not None:
+            self.config.update(self.initial_config)
+            self.config.apply_all()
+
         self.scattering_laser = 0
         self.fiber_led = 0
         self.top_led = 0
         self.state('standby')
+        self.clean_up_threads()
+        if len(self._threads):
+            self.logger.warning(f'There are {len(self._threads)} still alive in Arduino')
+        self.driver.close()
         super().finalize()
