+++ NanoCETPy/views/GUI/Measurement_Widget.ui
