+++ NanoCETPy/views/GUI/folder_icon.png
