+++ NanoCETPy/views/camera_viewer_widget.py
