--- cam_feed.yml
@@ -1,167 +1,0 @@
-%YAML 1.2
----
-# Default parameters for the Cam Feed program
-# All parameters can be changed to accommodate user needs.
-# All parameters can be changed at runtime with the appropriate config window
-user:
-  name: Jakob
-
-electronics:
-  arduino:
-    port: 'ASRL4::INSTR'
-    baud_rate: 115200
-  vertical_axis: 2
-  horizontal_axis: 1
-  focus_axis: 3
-  led_side: 0
-  led_top: 1
-  led_fiber: 2
-  led_power: 3
-  led_processing: 4
-  led_initialising: 5
-  led_ready: 6
-
-laser:
-  power: 0
-
-mirror:
-  speed: 1
-
-info:
-  folder: "C:\\Users\\Jakob Schroeder\\Code\\NanoCETPy\\recorded"
-  cartridge_number: 2001001
-  description: "Diffusion Movie"
-  filename_fiber: fiber_end_{cartridge_number}_{i}.npy
-  filename_laser: laser_position_{cartridge_number}_{i}.npy
-  filename_microscope: microscope_{cartridge_number}_{i}.npy
-  filename_movie: movie_{cartridge_number}_{i}.h5
-
-centroid:
-  laser_power: 30  # This is the laser power that will be used in order to be consistent with centroid extraction
-  exposure_time: 1ms
-  gain: 0.
-
-saving:
-  auto_save: False
-  auto_save_waterfall: True
-  directory:  C:\\Users\\Aquiles\\Data
-  filename_video: Video # Can be the same filename for video and photo
-  filename_photo: Snap
-  filename_tracks: Tracks
-  filename_waterfall: Waterfall
-  filename_trajectory: Trajectory
-  filename_log: Log
-  max_memory: 2500 # In megabytes
-
-GUI:
-  length_waterfall: 20 # Total length of the Waterfall (lines)
-  refresh_time: 50 # Refresh rate of the GUI (in ms)
-
-camera_fiber:
-  model: basler # Should be a python file in model/cameras
-  init: daA1280 # Initial arguments to pass when creating the camera
-  #extra_args: [extra, arguments] # Extra arguments that can be passed when constructing the model
-  model_camera: Dart # To keep a registry of which camera was used in the experiment
-  config:
-    exposure: 100ms # Initial exposure time (in ms)
-    gain: 14.
-    auto_exposure: Off
-    auto_gain: Off
-    pixel_format: Mono8
-    binning: [1, 1] # Binning
-    ROI: [[1, 1280], [1, 960]]
-    buffer_size: 200MB  # Buffer size to allocate memory in the ring-buffer of Basler
-
-camera_microscope:
-  model: basler # Should be a python file in model/cameras
-  init: a2A1920 #  acA1920 Initial arguments to pass when creating the camera
-  #extra_args: [extra, arguments] # Extra arguments that can be passed when constructing the model
-  model_camera: ACE # To keep a registry of which camera was used in the experiment
-  config:
-    exposure: 500us # Initial exposure time (in ms)
-    gain: 0.
-    ROI: [[1, 1920], [1, 1200]]
-    pixel_format: Mono12
-    auto_exposure: Off
-    auto_gain: Off
-    buffer_size: 2000MB
-
-camera_lumenera:
-  model: lumenera # Should be a python file in model/cameras
-  init: 22050065 # Initial arguments to pass when creating the camera, for lumenera using serial number seems straightforward 
-  #extra_args: [extra, arguments] # Extra arguments that can be passed when constructing the model
-  model_camera: None # To keep a registry of which camera was used in the experiment
-  config:
-    exposure: 500us # Initial exposure time (in ms)
-    gain: 1.
-    #auto_exposure: Off  # Not sure how to set this in Lucam API
-    #auto_gain: Off
-    pixel_format: 1 # PIXEL_FORMAT_16
-    binning: [1, 1] # Binning
-    ROI: [[1, 2064], [1, 1544]]
-    #buffer_size: 200MB  # Buffer size to allocate memory in the ring-buffer of Basler
-
-camera:
-  model: basler # Should be a python file in model/cameras
-  init: daA1280 # Initial arguments to pass when creating the camera
-  #extra_args: [extra, arguments] # Extra arguments that can be passed when constructing the model
-  model_camera: Dart # To keep a registry of which camera was used in the experiment
-  config:
-    exposure: 100ms # Initial exposure time (in ms)
-    gain: 1.
-    auto_exposure: Off
-    auto_gain: Off
-    pixel_format: Mono8
-    #binning: [1, 1] # Binning
-    ROI: [[1, 1280], [1, 960]]
-    buffer_size: 200MB  # Buffer size to allocate memory in the ring-buffer of Basler
-
-movie:
-  buffer_length: 5000 # Frames
-
-tracking:
-  locate:
-    diameter: 5  # Diameter of the particles (in pixels) to track, has to be an odd number
-    invert: False
-    minmass: 100
-  link:
-    memory: 3
-    search_range: 5
-  filter:  # Filter spurious trajectories
-    min_length: 25
-  process:
-    compute_drift: False
-    um_pixel: 0.15  # Microns per pixel (calibration of the microscope)
-    min_traj_length: 2
-    min_mass: 0.05
-    max_size: 50.0
-    max_ecc: 1
-    fps: 30
-  param_1: 0.
-  param_2: 0
-
-debug:
-  logging_level: Nothing # One of Nothing, Debug, Info, Warning, Error
-  queue_memory: False
-  to_screen: True
-
-sample:
-  name: a123
-  description: Short Description
-  cartrdige_number: 19NNMM
-
-laser_focusing:
-  high:
-    exposure_time: 400ms # originally 100ms
-    gain: 10.
-  low:
-    exposure_time: 10us
-    gain: 0.
-
-microscope_focusing:
-  high:
-    exposure_time: 50ms
-    gain: 10. # originally 40.
-  low:
-    exposure_time: 5ms
-    gain: 0.
